# Conflicts:
#	source/core/src/main/com/csse3200/game/areas/MainRoom.java

+++ source/core/src/main/com/csse3200/game/components/player/PlayerStatsDisplay.java
@@ -266,7 +266,7 @@ public class PlayerStatsDisplay extends UIComponent {
         speedProgressBar.setValue(speedPercentage);
     }
 
-    public void updateDamageUI(float damage) {
+    public void updateDamageUI(int damage) {
         damageProgressBar.setValue(damage);
     }
